Fix printf format for bytes read/written output.

+++ io/pread.c
@@ -196,8 +196,8 @@ pread_f(
 	gettimeofday(&t2, NULL);
 	t2 = tsub(t2, t1);
 
-	printf(_("read %ld/%ld bytes at offset %lld\n"),
-		(long)total, (long)count, (long long)offset);
+	printf(_("read %lld/%lld bytes at offset %lld\n"),
+		total, count, (long long)offset);
 	cvtstr((double)total, s1, sizeof(s1));
 	cvtstr(tdiv((double)total, t2), s2, sizeof(s2));
 	timestr(&t2, ts, sizeof(ts));
